+++ NEWS.md
@@ -1,5 +1,7 @@
 # od 0.2.1
 
+* Updated `od_disaggregate()` so it's less likely to fail with unfathomable message (#30)
+
 # od 0.2.0
 
 * New `od_split()` function, a copy of `od_disaggregate()`